CoreLibrary 	- fixed many warnings on visual c++ 2005 	- added pragma once (when msvc or gcc ) on often used and large include files 	  to speed up compilation (works very well with vc6+vc8)
Visual C++ project files
	- Replace /Fr by /FR in dsp project for more vc8 compatibility
	


git-svn-id: file:///fltk/svn/trunk@5119 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Clock.cxx
@@ -171,7 +171,7 @@ int Clock::handle(int event) {
     remove_timeout();
   case TIMEOUT:
 #ifdef _WIN32
-    value(time(0));
+    value((unsigned long) time(0));
     repeat_timeout(1.0f);
 #else
     {struct timeval t; gettimeofday(&t, 0);
